final/requirements.txt
SQLAlchemy==2.0.27
python-dotenv==1.0.1
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
networkx==3.2.1
matplotlib==3.8.2 